Add pre- and post- app loop actions. Add fool-protection for headers setting
These protections are:

1. If headers were not sent (SET_HTTP_STATUS_AND_HDR() were not called) then
   first APP_WRITE() call causes automatic SET_HTTP_STATUS_AND_HDR(200, NULL)
   Another subsequent APP_WRITE() calls will be just regular, and
   SET_HTTP_STATUS_AND_HDR() will cause nothing.
2. If SET_HTTP_STATUS_AND_HDR() were called first, then any further calls to
   SET_HTTP_STATUS_AND_HDR() will cause nothing. Also, APP_WRITE() will become
   regular immediately.

Also, these protections were made not only for protect app from missuse, but
for keeping consistency with requirements that were made with mongoose integration
and implementation. In other words, you are not able to use chunked http encoding
until you delcared it on headers, therefore NOT using SET_HTTP_STATUS_AND_HDR()
will cause just invalid data output, which is just bad.

These changes were made not with dumb if() checking, but with changing function
pointer values, which is the best way to deal with such situations.